requirements.txt
numpy~=1.24.4
pillow~=10.4.0
setuptools~=68.2.0
pytesseract~=0.3.10
opencv-python~=4.5.5.64
pandas~=2.0.3
PyMUPDF